+++ lib/bolt/bolt_option_parser.rb
@@ -363,14 +363,14 @@ module Bolt
         @options[:user] = user
       end
       define('-p', '--password [PASSWORD]',
-             'Password to authenticate with. Omit the value to prompt for the password.') do |password|
-        if password.nil?
-          STDOUT.print "Please enter your password: "
-          @options[:password] = STDIN.noecho(&:gets).chomp
-          STDOUT.puts
-        else
-          @options[:password] = password
+             'Password to authenticate with') do |password|
+        # TODO: Remove deprecation message
+        unless password
+          msg = "Optional parameter for --password is deprecated and no longer prompts for password. " \
+                "Use the prompt plugin instead to prompt for passwords."
+          raise Bolt::CLIError, msg
         end
+        @options[:password] = password
       end
       define('--private-key KEY', 'Private ssh key to authenticate with') do |key|
         @options[:'private-key'] = key
@@ -390,14 +390,14 @@ module Bolt
         @options[:'run-as'] = user
       end
       define('--sudo-password [PASSWORD]',
-             'Password for privilege escalation. Omit the value to prompt for the password.') do |password|
-        if password.nil?
-          STDOUT.print "Please enter your privilege escalation password: "
-          @options[:'sudo-password'] = STDIN.noecho(&:gets).chomp
-          STDOUT.puts
-        else
-          @options[:'sudo-password'] = password
+             'Password for privilege escalation') do |password|
+        # TODO: Remove deprecation message
+        unless password
+          msg = "Optional parameter for --sudo-password is deprecated and no longer prompts for password. " \
+                "Use the prompt plugin instead to prompt for passwords."
+          raise Bolt::CLIError, msg
         end
+        @options[:'sudo-password'] = password
       end
 
       separator "\nRun context:"